requirements.txt
numpy>=1.26.0
matplotlib>=3.8.0
pandas>=2.1.0
streamlit>=1.28.0
plotly>=5.17.0
reportlab>=4.0.0
pillow>=10.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
pytest>=7.4.0
pytest-cov>=4.1.0
requests>=2.31.0
python-multipart>=0.0.6
scikit-learn>=1.3.0
stripe>=7.0.0
httpx>=0.25.0
asyncpg>=0.29.0
sqlalchemy>=2.0.0
